+++ src/Optimisarr.Data/OptimisarrDbContext.cs
@@ -52,6 +52,7 @@ public sealed class OptimisarrDbContext(DbContextOptions<OptimisarrDbContext> op
             entity.HasIndex(file => file.Path).IsUnique();
             entity.Property(file => file.RelativePath).HasMaxLength(1024);
             entity.Property(file => file.Status).HasConversion<string>().HasMaxLength(32);
+            entity.Property(file => file.MediaKind).HasConversion<string>().HasMaxLength(32);
             entity.Property(file => file.Container).HasMaxLength(32);
             entity.Property(file => file.VideoCodec).HasMaxLength(64);
             entity.Property(file => file.AudioCodecs).HasMaxLength(256);